🔨 refactor: optionSelector위치 변
아토믹 디자인에 맞게 UI폴더에서 Modules 폴더로 변경
